feat: add Leetcode 73 solution
Set Matrix Zeroes
Practice for arrays. Traversed a 2D matrix while keeping the space
complexity constant. Used the first row and column as the indicators of
when a zero was present in that row or column. Needed a bool for row
zero since the first cell was for column zero.